requirements.txt
# Core dependencies
scapy==2.5.0
flask==3.0.0
shodan==1.30.1
python-nmap==0.7.1
netifaces==0.11.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
colorama==0.4.6
paramiko==3.4.0

# Additional security tools
mac-vendor-lookup==0.1.12
pyOpenSSL==24.0.0
cryptography==42.0.5
pycrypto==2.6.1
impacket==0.11.0
pypcap==1.3.0
netfilterqueue==1.0.0
python-libnmap==0.7.3

# Web interface enhancements
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
Flask-WTF==1.2.1

# Additional utilities
psutil==5.9.8
pysnmp==4.4.12
pexpect==4.9.0